Agent: Extract requests facade from HTTPIslandAPIClient
